pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "TortankWebServer"
version = "0.1.0"
authors = [
  	{ name="Trioen Loïc", email="trioen.loic@gmail.com" },
  	{ name="Mathys Deboever", email="" },
  	{ name="Samy", email="" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
	'Flask',
	'gpiozero',
	'ADS1x15-ADC'
]

[project.scripts]
TortankWebServer = "TortankWebServer.__main__:main"
CarapuceWebServer = "CarapuceWebServer.__main__:main"

[project.urls]
Repository = "https://github.com/AlphaGaming7780/TortankWebServer.git"
Homepage = "https://github.com/AlphaGaming7780/TortankWebServer"
Issues = "https://github.com/AlphaGaming7780/TortankWebServer/issues"